Migrate the themes components to align with Phlex conversions
