+++ app/components/themes/grid/calendar.rb
@@ -1,0 +1,14 @@
+module Components
+  module Themes
+    module Grid
+      class Calendar < Components::Base
+        def view_template
+          div(class: "space-y-4 w-full") do
+            Input(type: "string", placeholder: "Select a date", class: "rounded-md border shadow", id: "formatted-date", data_controller: "input")
+            Calendar(input_id: "#formatted-date", date_format: "PPPP", class: "rounded-md border shadow")
+          end
+        end
+      end
+    end
+  end
+end
